Resolve candidates using all available agent data
Extract _resolve_candidates() helper that AND's all non-empty fields
(name, set_code, collector_number, artist) per agent entry instead of
requiring set_code+collector_number. Fixes resolution when agent returns
empty collector_number but correct name/set/artist.

Replaces resolution logic in both _process_image_core and
_api_ingest2_update_cards.

Also moves TestWebImportResolve (Podman container test) from
tests/test_import.py to tests/integration/ where it belongs — it was
hanging the unit test suite on hosts with podman available.
